Move 'Blocked user link for modal' setting from 'General settings' to 'Fees settings'
