+++ web/modules/custom/dpl_react_apps/dpl_react_apps.module
@@ -376,9 +376,9 @@ function dpl_react_apps_dpl_react_apps_data(array &$data, array &$variables): vo
     $fees_settings->getViewFeesAndCompensationRatesUrl(),
   ];
 
-  if (!empty($general_settings->getBlockedPatronElinkUrl())) {
+  if (!empty($fees_settings->getBlockedPatronElinkUrl())) {
     $data['urls'] += [
-      'blocked-patron-e-link' => $general_settings->getBlockedPatronElinkUrl(),
+      'blocked-patron-e-link' => $fees_settings->getBlockedPatronElinkUrl(),
     ];
   }
 }
